refactor: restructure headers
