--- src/components/Layout/AppHeader/types.ts
@@ -1,3 +1,0 @@
-export interface AppHeaderProps {
-  backgroundColor?: string;
-}
