+++ src/components/Layout/Header/AvatarHeader/index.tsx
@@ -4,11 +4,12 @@ import { TouchableOpacity } from "react-native-gesture-handler";
 import { useAuth } from "~/contexts/auth/AuthContext";
 import useNavigate from "~/hooks/useNavigate";
 import { PROFILE_ROUTE } from "~/router/routes";
+import { HeaderContainer } from "~/components/Layout/Header/styles";
 
-import { HeaderContainer, HeaderUserAvatar } from "./styles";
-import { AppHeaderProps } from "./types";
+import { HeaderUserAvatar } from "./styles";
+import { AvatarHeaderProps } from "./types";
 
-const AppHeader: React.FC<AppHeaderProps> = ({ children, ...rest }) => {
+const AvatarHeader: React.FC<AvatarHeaderProps> = ({ children, ...rest }) => {
   const { user } = useAuth();
   const navigate = useNavigate();
 
@@ -23,4 +24,4 @@ const AppHeader: React.FC<AppHeaderProps> = ({ children, ...rest }) => {
   );
 };
 
-export default AppHeader;
+export default AvatarHeader;
