+++ src/components/Layout/Header/AvatarHeader/styles.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components/native";
+import Avatar from "~/components/Base/Avatar";
+
+export const HeaderUserAvatar = styled(Avatar)`
+  width: 56px;
+  height: 56px;
+  border-radius: 28px;
+`;
