+++ src/components/Layout/Header/AvatarHeader/types.ts
@@ -1,0 +1,5 @@
+import { ViewProps } from "react-native";
+
+import { HeaderContainerProps } from "~/components/Layout/Header/types";
+
+export type AvatarHeaderProps = ViewProps & Pick<HeaderContainerProps, "backgroundColor">
