+++ src/components/Layout/Header/styles.ts
@@ -1,14 +1,12 @@
 import styled from "styled-components/native";
 import { getStatusBarHeight } from "react-native-iphone-x-helper";
 
-import Avatar from "~/components/Base/Avatar";
+import { HeaderContainerProps } from "./types";
 
-import { AppHeaderProps } from "./types";
-
-export const HeaderContainer = styled.View<AppHeaderProps>`
+export const HeaderContainer = styled.View<HeaderContainerProps>`
   width: 100%;
   padding: 24px;
-  padding-top: ${getStatusBarHeight() + 24}px;
+  padding-top: ${getStatusBarHeight() + 20}px;
   flex-direction: row;
   align-items: center;
   justify-content: space-between;
@@ -16,10 +14,4 @@ export const HeaderContainer = styled.View<AppHeaderProps>`
     theme,
     backgroundColor,
   }) => backgroundColor || theme.colors.blackMedium};
-`;
-
-export const HeaderUserAvatar = styled(Avatar)`
-  width: 56px;
-  height: 56px;
-  border-radius: 28px;
 `;
