+++ src/components/Layout/Header/types.ts
@@ -1,0 +1,5 @@
+import { ViewProps } from "react-native";
+
+export interface HeaderContainerProps extends ViewProps {
+  backgroundColor?: string;
+}
